add pybind example with a struct
